+++ R/jsFunc-visibilityFuncs.R
@@ -6,6 +6,10 @@
 #' an element invisible, \strong{\code{toggle}} displays the element if it it
 #' hidden and hides it if it is visible.
 #'
+#' If \code{condition} is given to \code{toggle}, that condition will be used
+#' to determine if to show or hide the element. The element will be shown if the
+#' condition evalutes to \code{TRUE} and hidden otherwise.
+#'
 #' @param ... The following parameters are available:
 #' \tabular{ll}{
 #'   \strong{\code{id}}         \tab The id of the element/Shiny tag \cr
@@ -17,6 +21,8 @@
 #'   \strong{\code{time}}       \tab The number of seconds to make the
 #'                                   animation last
 #'                                   (default: \code{0.5}) \cr
+#'   \strong{\code{condition}}  \tab An optional argument to \code{toggle}, see
+#'                                   'Details' below. \cr
 #' }
 #' @seealso \code{\link[shinyjs]{useShinyjs}},
 #' \code{\link[shinyjs]{runExample}},
@@ -53,6 +59,23 @@
 #'   show(id = "element", anim = TRUE)
 #'   hide("element")
 #'   hide(id = "element", anim = TRUE)
+#' }
+#'
+#' ## toggle can be given an optional `condition` argument, which
+#' ## determines if to show or hide the element
+#' if (interactive()) {
+#'   shiny::shinyApp(
+#'     ui = shiny::fluidPage(
+#'       useShinyjs(),
+#'       shiny::checkboxInput("checkbox", "Show the text", TRUE),
+#'       shiny::p(id = "element", "Watch what happens to me")
+#'     ),
+#'     server = function(input, output, session) {
+#'       shiny::observe({
+#'         toggle(id = "element", condition = input$checkbox)
+#'       })
+#'     }
+#'   )
 #' }
 #' @name visibilityFuncs
 NULL